% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctheight.R
\name{correctheight}
\alias{correctheight}
\title{Correct height}
\usage{
correctheight(x)
}
\arguments{
\item{x}{A number.}
}
\value{
The sum of \code{x}
}
\description{
Correct height
}
\examples{
xvec <- c("160.cm","1,54","345 m")
openquestnum(xvec)




}
