% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editweight.R
\name{editweight}
\alias{editweight}
\title{Add together two numbers}
\usage{
editweight(x)
}
\arguments{
\item{x}{A number}

\item{y}{A number}
}
\value{
The sum of \code{x} and \code{y}
}
\description{
Add together two numbers
}
\examples{
add(1, 1)
add(10, 1)
}

