% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractsechtml.R
\name{extractsechtml}
\alias{extractsechtml}
\title{Extract Sections from html}
\usage{
extractsechtml(htmldat)
}
\arguments{
\item{htmldat}{An html document.}
}
\value{
The sections of \code{htmldat}
}
\description{
Extract Sections from html
}
\examples{

}
