% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openquestfreq.R
\name{openquestfreq}
\alias{openquestfreq}
\title{Transform Open Question about Frequency}
\usage{
openquestfreq(x)
}
\arguments{
\item{x}{A frequency.}
}
\value{
The frequency of \code{x}
}
\description{
Transform Open Question about Frequency
}
\examples{
xvec <- c("ca. 2","20","-77","0.5")
openquestfreq(xvec)

x <- c("ca.4-6","-77","2-3 x","1 mal","2 mal","ein","zwei","Nie","sehr oft / ca. 20 mal")
openquestfreq(x)

x <- c("4,5","4.5","nein","einmal","ein halbes Jahr","ja","JA","02.3.")
openquestfreq(x)

}
