% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openquestnum.R
\name{openquestnum}
\alias{openquestnum}
\title{Transform Open Question}
\usage{
openquestnum(x)
}
\arguments{
\item{x}{A number.}
}
\value{
The sum of \code{x}
}
\description{
Transform Open Question
}
\examples{
xvec <- c("160.cm","1,54","345 m")
openquestnum(xvec)




}
